Split LLMPipeline by several files (#1454)
