+++ src/cpp/src/llm_pipeline_base.hpp
@@ -13,8 +13,26 @@ namespace genai {
 class LLMPipelineImplBase {
 public:
     LLMPipelineImplBase(const Tokenizer& tokenizer,
-                        const GenerationConfig& config = {})
-    : m_tokenizer(tokenizer), m_generation_config(config) {
+                        const GenerationConfig& config)
+    : m_tokenizer(tokenizer), m_generation_config(config) { }
+
+    Tokenizer get_tokenizer() {
+        return m_tokenizer;
+    }
+
+    GenerationConfig get_generation_config() const {
+        return m_generation_config;
+    }
+
+    void set_generation_config(GenerationConfig config) {
+        int64_t default_eos_token_id = m_generation_config.eos_token_id;
+        m_generation_config = config;
+
+        // if eos_token_id was not provided in config forward from default config
+        if (m_generation_config.eos_token_id == -1)
+            m_generation_config.set_eos_token_id(default_eos_token_id);
+
+        m_generation_config.validate();
     }
 
     virtual DecodedResults generate(
@@ -34,6 +52,12 @@ public:
 
     virtual ~LLMPipelineImplBase() = default;
 
+    void save_load_time(std::chrono::steady_clock::time_point start_time) {
+        auto stop_time = std::chrono::steady_clock::now();
+        m_load_time_ms = std::chrono::duration_cast<std::chrono::milliseconds>(stop_time - start_time).count();
+    }
+
+protected:
     Tokenizer m_tokenizer;
     GenerationConfig m_generation_config;
     std::optional<AdapterController> m_adapter_controller;
